+++ src/ambition/version_details.cc
@@ -22,6 +22,7 @@ std::vector<VersionUpdates> versionDetails = {
           _("Show building construction costs on buttons."),
           _("Show war machine and ship construction costs on War Factory and Harbour build menus."),
           _("Show total production queue count and cost for War Factories and Harbours."),
+          _("Show keybinds at the bottom-left of buttons that have them."),
         },
       },
       {