+++ CHANGELOG.md
@@ -16,6 +16,7 @@ speaking, considered to be the player's experience.
 - Show war machine and ship construction costs on War Factory and Harbour build
   menus.
 - Show total production queue count and cost for War Factories and Harbours.
+- Show keybinds at the bottom-left of buttons that have them.
 
 ### Changed